backend/requirements.txt
# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter
python-dotenv>=1.0.0    # For loading .env files
alembic>=1.12.0         # For database migrations

# FastAPI
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# LangGraph
langgraph>=0.0.15

# Utilities
python-jose[cryptography]>=3.3.0  # For JWT tokens
passlib[bcrypt]>=1.7.4           # For password hashing
python-multipart>=0.0.5          # For form data processing
email-validator>=2.0.0           # For email validation